+++ lib/doorkeeper/config.rb
@@ -405,6 +405,11 @@ module Doorkeeper
       @token_grant_types ||= calculate_token_grant_types.freeze
     end
 
+    def allow_blank_redirect_uri?
+      grant_flows.exclude?("authorization_code") &&
+        grant_flows.exclude?("implicit")
+    end
+
     def option_defined?(name)
       !instance_variable_get("@#{name}").nil?
     end