+++ lib/generators/doorkeeper/migration_generator.rb
@@ -4,6 +4,8 @@ require "rails/generators"
 require "rails/generators/active_record"
 
 module Doorkeeper
+  # Copies main Doorkeeper migration into parent Rails application.
+  #
   class MigrationGenerator < ::Rails::Generators::Base
     include ::Rails::Generators::Migration
     source_root File.expand_path("templates", __dir__)