+++ spec/models/doorkeeper/application_spec.rb
@@ -82,10 +82,34 @@ module Doorkeeper
       expect(new_application).not_to be_valid
     end
 
-    it "is invalid without redirect_uri" do
-      new_application.save
-      new_application.redirect_uri = nil
-      expect(new_application).not_to be_valid
+    context "redirect URI" do
+      context "when grant flows allow blank redirect URI" do
+        before do
+          Doorkeeper.configure do
+            grant_flows %w[password client_credentials]
+          end
+        end
+
+        it "is valid without redirect_uri" do
+          new_application.save
+          new_application.redirect_uri = nil
+          expect(new_application).to be_valid
+        end
+      end
+
+      context "when grant flows require redirect URI" do
+        before do
+          Doorkeeper.configure do
+            grant_flows %w[password client_credentials authorization_code]
+          end
+        end
+
+        it "is invalid without redirect_uri" do
+          new_application.save
+          new_application.redirect_uri = nil
+          expect(new_application).not_to be_valid
+        end
+      end
     end
 
     it "checks uniqueness of uid" do